requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-extensions==3.2.3
channels==4.2.0
channels-redis==4.2.1
firebase-admin==6.5.0
django-environ==0.12.0
gunicorn==23.0.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
redis==5.0.1
librosa==0.10.2.post1
numpy==2.0.2
soundfile==0.12.1
webrtcvad==2.0.10
stripe==7.8.0 